Create a test base superclass.
