+++ tests/views/view_tests_base.py
@@ -1,32 +1,13 @@
 """Tools to help run tests against the Django app."""
 
-import os
-import unittest
-
-import django
-import django.test
-from google.appengine.ext import testbed
-
 import const
 
 import scrape
+import testutils.base
 
 
-class ViewTestsBase(unittest.TestCase):
+class ViewTestsBase(testutils.base.ServerTestsBase):
     """A base class for tests for the Django app."""
-
-    def setUp(self):
-        self.testbed = testbed.Testbed()
-        self.testbed.activate()
-        self.testbed.init_user_stub()
-        os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'settings')
-        django.setup()
-        django.test.utils.setup_test_environment()
-        self.client = django.test.Client()
-
-    def tearDown(self):
-        self.testbed.deactivate()
-        django.test.utils.teardown_test_environment()
 
     def login(self, is_admin=False):
         """Logs in the "user" for making requests.
